pipeline/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
yt-dlp
openai-whisper
torch
